refactor: re-organize all components into a better hierarchy
