+++ interchain-ui/src/components/Navbar.tsx
@@ -1,0 +1,26 @@
+import { ConnectWalletButton } from '@agoric/react-components';
+
+const Navbar = () => {
+  return (
+    <div className="daisyui-navbar bg-neutral text-neutral-content">
+      {/* Agoric logo */}
+      <div className="flex-none">
+        <button className="daisyui-btn daisyui-btn-square daisyui-btn-ghost">
+          <img src="/agoric.svg" />
+        </button>
+      </div>
+      {/* dApp title */}
+      <div className="flex-1">
+        <button className="daisyui-btn daisyui-btn-ghost text-xl">
+          dApp Agoric Basics
+        </button>
+      </div>
+      {/* connect wallet button */}
+      <div className="flex-none">
+        <ConnectWalletButton className="daisyui-btn daisyui-btn-outline daisyui-btn-secondary" />
+      </div>
+    </div>
+  );
+}
+
+export {Navbar};
